requirements.txt
uagents>=0.6.0
agentverse
langchain>=0.1.0
langchain-community>=0.0.10
pydantic>=2.0.0
openai>=0.27.0
spotipy>=2.23.0
requests
streamlit>=1.18.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.6 